apps/models/pyproject.toml
[project]
name = "models"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "datasets>=3.5.0",
    "fastapi[standard]>=0.112.1",
    "huggingface-hub>=0.30.2",
    "peft>=0.15.2",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
    "transformers>=4.51.3",
    "validators>=0.34.0",
]
